Added MCMCIterators as a dependency for the project. Small update to README.
Added a sampler to the mcmc capabilities that just considers a fixed sample (i.e., delta function distribution).

Added a magnetic field scaling parameter (exposed in HallThruster.jl) as an *optional* operating condition parameter. When omitted, assumes unity.

Modified cathode flow fraction definition to be consistent with literature for the SPT-100 and the SPT-140 (and more broadly conventional).

Hot fix to solve cases where relative error calculations produce negative error (because corresponding data are negative).
Note that this is still poor behavior. Really assuming relative error of any sort is ill-posed in this case. But this stops the code from failing, at least.